Document common Conventions and Steps
When I started using Cake.Recipe, I came up with a lot of questions and little things that prevented me from using it. Essentially, it was an understanding of current conventions and how to tweak them along with a perspective of the problem at hand and how to solve with this tool. Adding this note, to hopefully help the next person who comes along to learn.. This needs feedback