(GH-535) Provide ability to not use Coveralls
Previously, it was assumed that you would always publish to Coveralls,
and the only way to prevent it was to hack the environment variable to
be something that couldn't be found.  This new parameter allows
controlling whether Coveralls should be used.  It will default to true,
since that is how things work just now.