(GH-523) Pin to specific version of coveralls
The coveralls.net NuGet package is being used for both the full .net
framework version and the .Net Global tool version.  Pin to 0.7.0 for
the latest .Net framework version (which will mean it will run on Unix
systems) and also update to latest Cake.Coveralls addin which includes
a fix for a case sensitive issue.