(maint) Set branch name to empty string in case one was not found
In some edge cases the branch result from one of the
Build Providers will be null and cause a null reference
exception when we try to compare that branch value with
a different value. This commit changes this to always
defaulting to an empty string when the result returned
is null.