(maint) Ensure there are any files before we try to copy them
In some cases there may be multiple output paths,
but 1 or more of the paths do not contain any files.
This typically happens if conditional target frameworks
are being used in the project file. In these cases
trying to call 'CopyFiles' results in an exception being
thrown. This commit adds an additonal check before
calling 'CopyFiles' to ensure we aren't passing in an empty
enumerable.