(GH-531) Improve branch detection on github actions
During a normal branch build on GitHub actions, the branch reference
is prefixed with 'refs/heads' and should be removed from reference
to get the correct branch name. This commit makes the change
necessary by checking the tempName if it starts with that string
and remove it if it exist.