(GH-628) Copy system libgit2 files to cake.git and gitversion directories
On some linux distros it is not possible to use
the embedded libgit2 binaries for these two
softwares, and as such they need to either be linked
or be copied to the necessary directories.
As we are moving toward better support regarding
.NET Core cake, we can no longer fully provide
the support of patch DLLMaps as these do not
work using .NET Core. As such this commit makes the
necessary changes to copy the system libgit2 library to the existing
libgit2*.so locactions underneeth the tools directory.