(GH-700) renamed GITHUB_TOKEN to GITHUB_PAT
to ensure no conflicts between gh-cli and Cake.Recipe.
Also, when evaluating the GITHUB_PAT the following
environment-variables are considered (only if GITHUB_PAT is empty):
* GH_TOKEN
* GITHUB_TOKEN

This ensures backwards compatibility with previous versions of Cake.Recipe
and also conforms a bit more to gh-cli, as the two are used there, too.