(GH-768) Lookup branch name from git
This is similar to what we have had to do in the GitHub Actions
provider, since the branch name is not part of the information that is
provided directly. When the branch name contains refs/tags, execute git
to find the branch name, otherwise, accept the branch name that is
provided.