(build) Revert back to earlier version of Cake
Initially, I thought that I would "have" to update this, in order to
start using the Figlet change that I added to the recipe, however, this
type of change is only required in the "parent" project that is using
Cake.Recipe. This change will be made once we release the next version
of Cake.Recipe.