(#897) Added a configuration for the MessageParameters
So now, when StandardMessage, TwitterMessage, GitterMessage,
or MicrosoftTeamsMessage are changed, the parameters
which are send to string.Format can also be changed by
settings MessageParameters.

The default for MessageParameters is equal to the "old" value,
so this change is compatible with "old" settings where the
messages were either changed or unchanged.