(#921) Read contents of yml file and pass to MSBuild
A new cake-version.yml file is added to the project structure and includes
the current target version of Cake to be used when using Cake.Recipe.

To not have the version in multiple places, the yml is read during the
build and the cake-version from the yml is passed to MSBuild using
a property.