(#992) Provide ability to set ignore of KuduSync
Rather than hard-coding the ignore list, here we have added a
ToolSettings property, which will allow this ignore list to be set
within the recipe.cake file, to control exactly what files are ignored
during a sync.

The default value of the ignore list remains the same as it was
previously.