(#1019) Baseline update to Cake 3.0.0
Cake 3.0.0 had a number of breaking changes, the main one being the
removal of some obsoleted aliases and members;

https://cakebuild.net/blog/2022/11/cake-v3.0.0-released#breaking-changes

We need to remove the usage of these across the Cake script within
Cake.Recipe.

In addition, bumped to the updated versions of some Cake Addins, to
allow the top level build to succeed.