(#1078) Remove usage of Wyam
Wyam is no longer supported, and there are a number of errors when
trying to execute it as part of CI builds.  Rather than fight it, the
decision has been made to remove the usage of it for just now, and
replace is with "something" later on.  The top level documentation
tasks have been left in place, and will be filled when a new tool is
selected. Due to the fact that the publishing of the documentation was
the only place that the Kudu tool was being used, this has also been
removed.